+++ wiki/Changelog.md
@@ -5,6 +5,7 @@ Terminal features
   * Tweak DECRQCRA Rectangular Checksum to match xterm (~#1225, xterm 390).
   * Fix status area refresh, especially after restoring a minimized window.
   * Workaround for rendering failure of top-left background text on background image.
+  * Fix invisible and blinking attributes on background image.
 
 Misc
   * Fix handling of options to be saved (#1246, #1247).